requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.1
pandas>=1.5.0
google-cloud-bigquery>=3.5.0
apscheduler>=3.9.1
icalendar>=6.3.1
gcloud
selenium
webdriver-manager
playwright
gspread